Extracted PlayerSlot into (PlayerColor and PlayerSlot)
